Allow all RFC3986-compliant URLs
This expands the URL syntax to allow all RFC3986-compliant URLs.  This
means that, in particular:

 * empty path segments are now permitted
 * an entirely empty path is also permitted (though it is defined to be
   equivalent to the path `/`, which has a single empty segment)
 * paths may contain the characters '(', ')' and ','
   * strictly, any path segment must be entirely `pchar`s or
     `path-characters`, so some combinations are still disallowed, such
     as "foo,bar^baz" because ',' is not a `path-character` and '^' is
     not a `pchar`.

As this purely expands the space of strings which can be parsed as URLs,
this is a mostly-not-breaking change; no existing URLs will fail to
parse.  However, there are some dhall programs whose interpretation will
change, because a URL will consume more characters than it previously
was able to.  For example

    http://example.com/foo//bar

was previously parsed as

    http://example.com/foo ⫽ bar

but is now parsed as a single URL.  Another example is

    [ http://example.com/foo, bar ]

which previously was a list containing two items
`http://example.com/foo` and `bar`, and is now a list with a single
item: the result of applying `http://example.com/foo,` to `bar`.

A final awkward example is:

    (http://example.com/foo)

which previously parsed as `http://example.com/foo`, but now fails to
parse (in a PEG grammar, at least) because the final ')' character is
greedily consumed by the URL parser and then we have nothing to close
the leading parenthesis.

The last two examples could be avoided if we disallow the characters
'(', ')' and ',' from the `pchar` rule.  This would get us closer to
standard URL syntax but avoid some tricky pitfalls of parsing.  My
preference is to go with the standard `pchar` rule despite the edge
cases.

Fixes #205.  See also #581.